writing and reading huffman encoding

+++ Compression/Leaf.cpp
@@ -4,6 +4,6 @@
 
 #include "Leaf.h"
 
-Leaf::Leaf(int freq, int c) : FNode(freq) {
+Leaf::Leaf(int freq, uint8_t c) : FNode(freq) {
     color = c;
 }
